Fix type safety issues with strlen() return values
- Use size_t instead of int for string length variables in html2.c and mystrings.c
- Add explicit cast to long for strlen() result in cl_generic.c
- Simplify deBracket() function by removing unused variables and inlining strlen()

+++ src/viola/cl_generic.c
@@ -4324,7 +4324,7 @@ long meth_generic_str(VObj* self, Packet* result, int argc, Packet argv[]) {
 long meth_generic_strlen(VObj* self, Packet* result, int argc, Packet argv[]) {
     clearPacket(result);
     result->type = PKT_INT;
-    result->info.i = strlen(PkInfos2Str(argc, argv));
+    result->info.i = (long)strlen(PkInfos2Str(argc, argv));
     return 1;
 }
